+++ src/components/ids-dropdown/ids-dropdown-list.ts
@@ -148,7 +148,7 @@ export default class IdsDropdownList extends Base {
 
         if (target) {
           this.lastHovered = (target as IdsListBoxOption);
-          this.dropdownEl?.selectTooltip(this.lastHovered);
+          this.dropdownEl?.selectTooltip?.(this.lastHovered);
         }
       });
     }
@@ -316,6 +316,13 @@ export default class IdsDropdownList extends Base {
       if (this.listBox.maxHeight && this.popup) {
         this.popup.maxHeight = this.listBox.maxHeight;
       }
+    }
+
+    if (this.listBox?.options?.length) {
+      this.listBox.options.forEach((option: IdsListBoxOption, index) => {
+        if (option.rowIndex >= 0) return;
+        option.rowIndex = index; // row-index used to keep track of the sort-order of options
+      });
     }
   }